Fix circular dependency on encoders
